Inicio do tratamento das exceptions

+++ src/main/java/desafio3/devsuperior/service/ClientService.java
@@ -1,16 +1,19 @@
 package desafio3.devsuperior.service;
 
 import desafio3.devsuperior.dto.ClientDto;
+import desafio3.devsuperior.exception.DatabaseIntegrityViolationException;
+import desafio3.devsuperior.exception.ResourceNotFoundException;
 import desafio3.devsuperior.model.Client;
 import desafio3.devsuperior.repository.ClientRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
-import java.util.List;
 
 @Service
 public class ClientService {
@@ -19,7 +22,7 @@ public class ClientService {
 
     @Transactional(readOnly = true)
     public ClientDto findById(Long id) {
-        Client client = clientRepository.findById(id).orElseThrow();
+        Client client = clientRepository.findById(id).orElseThrow(() -> new ResourceNotFoundException("Recurso não encontrado!"));
         return new ClientDto(client);
     }
 
@@ -40,14 +43,27 @@ public class ClientService {
 
     @Transactional
     public ClientDto update(Long id, ClientDto clientDto) {
-        Client client = clientRepository.getReferenceById(id);
-        transformDtoToEntity(clientDto, client);
-        return new ClientDto(client);
+        try {
+            Client client = clientRepository.getReferenceById(id);
+            transformDtoToEntity(clientDto, client);
+            return new ClientDto(client);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+
     }
 
     @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id) {
-        clientRepository.deleteById(id);
+        if (!clientRepository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado!");
+        }
+        try {
+            clientRepository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseIntegrityViolationException("Violação da integridade referencial!");
+        }
+
     }
 
     /*Método usado para converter de dto para entidade:*/
